ns for fx: move node I/O dtype mapping to be local instead of global
Summary:

To support shadows of custom functions, we need to allow user to
specify I/O type of the custom functions.

This PR is a cleanup in preparation for making the above happen.
We make the I/O dtype mappings be generated by a function instead
of a global variable. In the next PR, we will add a hook so user
can modify these mappings.

Test Plan:

```
python test/test_quantization.py TestFXNumericSuiteCoreAPIs
```

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 479a392ec683b9b952d3e90de8ca385b515bc834
Pull Request resolved: https://github.com/pytorch/pytorch/pull/56296